refactor: Rewrite health check server
